pydantic-graphs-101/pyproject.toml
[project]
name = "pydantic-graphs-101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.3.1",
    "pydantic-ai-slim[openai]>=0.0.19",
    "pydantic-graph>=0.0.19",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
]
